Add new thread management system calls and update syscall vectors
- Replaced the reserved syscall at index 0x186 with sys_p_sleepthread, allowing threads to sleep for a specified duration.
- Introduced new syscalls for thread management: sys_p_thread_sigmask, sys_p_thread_signal_mode, sys_p_kill_thread, sys_p_thread_handler, and sys_p_thread_sigwait, replacing previously reserved entries from 0x18b to 0x18f.
- Updated syscall master definitions to reflect the new thread management functions and their signatures.
