+++ sys/syscall_vectors.c
@@ -522,16 +522,16 @@ Func dos_tab [DOS_MAX] =
 	/* 0x183 */		sys_f_dirfd,	/* 1.17 */
 	/* 0x184 */		sys_enosys,		/* reserved */
 	/* 0x185 */		sys_p_yieldthread,
-	/* 0x186 */		sys_enosys,		/* reserved */
+	/* 0x186 */		sys_p_sleepthread,		/* reserved */
 	/* 0x187 */		sys_enosys,		/* reserved */
 	/* 0x188 */		sys_enosys,		/* reserved */
 	/* 0x189 */		sys_p_createthread,
 	/* 0x18a */		sys_p_exitthread,
-	/* 0x18b */		sys_enosys,		/* reserved */
-	/* 0x18c */		sys_enosys,		/* reserved */
-	/* 0x18d */		sys_enosys,		/* reserved */
-	/* 0x18e */		sys_enosys,		/* reserved */
-	/* 0x18f */		sys_enosys,		/* reserved */
+	/* 0x18b */		sys_p_thread_sigmask,
+	/* 0x18c */		sys_p_thread_signal_mode,
+	/* 0x18d */		sys_p_kill_thread,
+	/* 0x18e */		sys_p_thread_handler,
+	/* 0x18f */		sys_p_thread_sigwait,
 
 	/* 0x190 */		/* DOS_MAX */
 };
